+++ app/controllers/companies/configurations_controller.rb
@@ -15,7 +15,7 @@ module Companies
       service_call = Companies::Configurations::UpdateService.call(
         company: @company,
         params: configuration_params,
-        use_work_time: params[:jsonb_columns_configuration][:use_work_time] == '1'
+        use_work_time: ActiveModel::Type::Boolean.new.cast(params[:jsonb_columns_configuration][:use_work_time])
       )
       if service_call.success?
         redirect_to companies_path, notice: "Configuration for company #{@company.title} is updated"