+++ db/migrate/20230606135138_add_worktime_to_users.rb
@@ -1,0 +1,6 @@
+class AddWorktimeToUsers < ActiveRecord::Migration[7.0]
+  def change
+    add_column :users, :work_start_time, :time
+    add_column :users, :work_end_time, :time
+  end
+end